refactor: NumGenerator 클래스를 Model에서 Utils로 옮기다
